moved DEBUG.H to SRC for consistency
