+++ platformio/FujiNet/lib/hardware/fnUART.h
@@ -67,9 +67,9 @@ public:
 };
 
 // Only define these if the default Arduino global SerialX objects aren't declared
-//#ifdef NO_GLOBAL_SERIAL
+#ifdef NO_GLOBAL_SERIAL
 extern UARTManager fnUartDebug;
 extern UARTManager fnUartSIO;
-//#endif
+#endif
 
 #endif //FNUART_H
