+++ platformio/FujiNet/lib/printer-emulator/png_printer.cpp
@@ -1,5 +1,5 @@
 #include "png_printer.h"
-#include "debug.h"
+#include "../../src/debug.h"
 
 // rewrite of TinyPngOut https://www.nayuki.io/page/tiny-png-output
