+++ platformio/FujiNet/platformio.ini
@@ -30,7 +30,7 @@ framework = arduino
 build_flags =
     -D JTAG
     -D ESP_32
-    -D DEBUG_FN
+    -D DEBUG_S
     -D BUG_UART=Serial
     -D DEBUG_SPEED=921600
     -D NO_GLOBAL_SERIAL
@@ -40,8 +40,8 @@ build_flags =
     ;-mfix-esp32-psram-cache-issue
 ;board_build.partitions = fujinet_partitions_16MB.csv
 board_build.partitions = fujinet_partitions_4MB.csv
-upload_port = COM5 ;COM3
+upload_port = COM3 ;COM3
 upload_speed = 921600
-monitor_port = COM5 ;COM3
+monitor_port = COM3 ;COM3
 monitor_speed = 921600
 lib_ldf_mode = deep+
